Added a test and another function
Added Render_Image_Russian to render an image with Russian Roulette
without multithreading. This is mostly a copy-paste job with some
changes.

ValidateProject can now be compiled as seen in the Makefile. This
compiles the source file with the flag -DVALIDATE, this flag runs code
in the source file that validates rendering without Russian Roulette.