Fixed a long-standing bug.
The bug was in the function Hemisphere_Vector_Generator, the
if-statement checking the magnitude of a psuedo-random vector was
between 0 and 1 was flawed.

The method of creating a random value in SceneRender has been fixed.
A new seed, generation and distribution is only created once per
instance of SceneRender.

Created a validation case for the renderer with Russian Roulette
implemented.