Added Render_Image function + others.
The Render_Image function inside RenderSceneSDF required few modifications from the original, simply to add parameters.

Added the functions between the SDF Operations comment blocks and the SDF Functions comment blocks.
The SDF functions include the SDFs for a sphere, a cube and a torus.
The SDF operations include operations for rotating, displacing, taking the union, intersection and subtraction of SDFs.