Added MATLAB files for plotting results.
Added a virtual destructor to The PhysicalObject base class.

Fixed a bug with the Ray_To_PixelXY function in the Observer class regarding resolution. A division by zero occured whenever one of the resolution dimensions was 1.

Added a new parameter to Render_Image in SceneRender and SceneRenderSDF. This boolean allows the function to run without printing progress.

Added another part to the validation case that finds the variance of the Monte-Carlo method with a fixed number of bounces.

Driver code changes are inconsequential.