Added MATLAB files.
Edited the code with a commented block that may be uncommented to output images with only contributions from the n-th bounce.

Added MATLAB files that are used to plot the different values of variance and radiance in the validation case.

Added a new line to the Makefile that runs the code with optimisations.