Added SDF Classes.
No support for SDF classes yet in SceneRender. Work in progress.

The virtual base class SDF has been created, SDFs will have individual BRDFs and Light_Emitted functions.
There is also a function called Inverse_Transformation, defining this via a function pointer will transform the SDF by the inverse of Inverse_Transformation.

Added the SphereSDF class.

Added functionality for individual SDFs to be added to SceneRender, the function SceneRender::SDF_Fct returns the minimum of the SDFs supplied and a version of this function allows the index of the closest SDF object to be found.