Added more SDF functionality to SceneRender.
Implemented a new Light_Out function that works with SDFs.

The SDF constructors now require the boolean argument for whether the SDF is inverted or not.

The SphereSDF SDF function has been changed to switch signs if invert_SDF is true.

Fixed SDF_Fct in SceneRender by not allowing sdf_object_pt_vector to be accessed if it is empty.

Other small bug fixes.